+++ packages/viz/src/index.js
@@ -41,12 +41,7 @@ import DataUtil from './utils/data'
 import getParametersChanges from './utils/parametersHistory'
 import createPrintPDFPackage from './modules/print'
 import { generatePDFStats, generatePumpSettings, selectDailyViewData } from './utils/print/data'
-import {
-  delayShowCbgTracesOnFocus,
-  focusTrendsCbgDateTrace,
-  unfocusTrendsCbgDateTrace,
-  unfocusTrendsCbgSlice
-} from './redux/actions/index'
+import { delayShowCbgTracesOnFocus, unfocusTrendsCbgSlice } from './redux/actions/index'
 
 const components = {
   WarmUpTooltip
@@ -101,7 +96,5 @@ export {
   getLongDayHourFormat,
   formatParameterValue,
   delayShowCbgTracesOnFocus,
-  focusTrendsCbgDateTrace,
-  unfocusTrendsCbgDateTrace,
   unfocusTrendsCbgSlice
 }